requirements.txt
Pillow
numpy
matplotlib
scikit-image
